apps/backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
cachetools==6.1.0
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.2
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.2.3
drf-spectacular==0.29.0
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.30.0
social-auth-app-django==5.6.0
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
